Implement robust Huber/M-regression for DL and GBM: a hybrid of squared and linear loss, with a transition point in absolute error, here called 'huber_delta'.
It's important to set huber_delta to a meaningful value as regression target
values may vary greatly.

The user specifies huber_alpha, the quantile of the absolute error (e.g. 0.9
for 90-th percentile), so we first need to know how the model's predictions are
distributed before we can separate the outliers from the inliers and compute a
meaningful Huber loss.

The mean_residual_deviance for distribution="huber" is the Huber loss, and it
depends on huber_alpha (and the model and the data, obviously).

For GBM, the huber_delta will be re-computed for every tree during training, and for every test set during scoring.
For DL, the huber_delta will be re-computed after every scoring on the training set during training, and for every test set during scoring.